+++ src/test/java/test/com/qiniu/storage/FormUploadTest.java
@@ -14,6 +14,7 @@ import test.com.qiniu.TestConfig;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
+import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.CountDownLatch;
@@ -450,6 +451,47 @@ public class FormUploadTest {
             }
         }
     }
+    
+	/**
+	 * 测试inputStream 表单上传
+	 * 检测reqid是否为Null
+	 * 检测状态码是否为200
+	 */
+	@Test
+	public void testFormUploadWithInputStream() {
+		testFormUploadWithInputStream(1, -1);
+		testFormUploadWithInputStream(1, 0);
+		testFormUploadWithInputStream(1, 1000);
+		testFormUploadWithInputStream(4 * 1024, 4 * 1024 * 1024);
+		testFormUploadWithInputStream(5 * 1024, -1);
+		testFormUploadWithInputStream(5 * 1024, 5 * 1024 * 1024);
+	}
+	
+	/**
+	 * 测试inputStream 表单上传
+	 * 检测reqid是否为Null
+	 * 检测状态码是否为200
+	 */
+	public void testFormUploadWithInputStream(long kiloSize, long size) {
+		
+		String token = TestConfig.testAuth.uploadToken(TestConfig.testBucket_z0, TestConfig.testBucket_z0, 3600, null);
+		System.out.println("token="+token);
+		
+		try {
+			File file = TempFile.createFile(kiloSize);
+			InputStream inputStream = new FileInputStream(file);
+			System.out.println("length=" + file.length());
+			System.out.println("size=" + size);
+			Response response = uploadManager.put(inputStream, size, TestConfig.testBucket_z0, token, null, null, false);
+			System.out.println("code="+response.statusCode);
+			System.out.println("reqid="+response.reqId);
+			System.out.println(response.bodyString());
+			assertNotNull(response.reqId);
+			assertEquals(200, response.statusCode);
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+	}
 
     class MyRet {
         public String hash;
